requirements.txt
aiohttp==3.9.5
aiosignal==1.3.1
altair==5.3.0
annotated-types==0.7.0
anyio==4.4.0
appnope==0.1.4
arxiv==2.1.3
asttokens==2.4.1
attrs==23.2.0
backoff==2.2.1
beautifulsoup4==4.12.3
blinker==1.8.2
cachetools==5.4.0
certifi==2024.7.4
cffi==1.16.0
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.7
comm==0.2.2
cryptography==43.0.0
dataclasses-json==0.6.7
debugpy==1.6.7
decorator==5.1.1
deepdiff==7.0.1
distro==1.9.0
emoji==2.12.1
exceptiongroup==1.2.2
executing==2.0.1
faiss-cpu==1.8.0.post1
filetype==1.2.0
frozenlist==1.4.1
gitdb==4.0.11
GitPython==3.1.43
googleapis-common-protos==1.63.2
grpcio==1.65.1
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.7
importlib_metadata==8.1.0
ipykernel==6.29.5
ipython==8.26.0
jedi==0.19.1
Jinja2==3.1.4
joblib==1.4.2
JPype1==1.5.0
jsonpatch==1.33
jsonpath-python==1.0.6
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
jupyter_client==8.6.2
jupyter_core==5.7.2
kiwipiepy==0.18.0
kiwipiepy-model==0.18.0
konlpy==0.6.0
langchain==0.2.11
langchain-community==0.2.10
langchain-core==0.2.23
langchain-experimental==0.0.63
langchain-openai==0.1.17
langchain-teddynote==0.0.13
langchain-text-splitters==0.2.2
langchainhub==0.1.20
langdetect==1.0.9
langsmith==0.1.93
libmagic==1.0
lxml==5.2.2
lz4==4.3.3
markdown-it-py==3.0.0
MarkupSafe==2.1.5
marshmallow==3.21.3
matplotlib-inline==0.1.7
mdurl==0.1.2
mmh3==4.1.0
multidict==6.0.5
mypy-extensions==1.0.0
nest_asyncio==1.6.0
nltk==3.8.1
numpy==1.26.4
openai==1.37.0
ordered-set==4.1.0
orjson==3.10.6
packaging==24.1
pandas==2.2.2
parso==0.8.4
pdfminer.six==20231228
pdfplumber==0.11.2
pexpect==4.9.0
pickleshare==0.7.5
pillow==10.4.0
pinecone-client==5.0.0
pinecone-plugin-inference==1.0.2
pinecone-plugin-interface==0.0.7
pinecone-text==0.9.0
pip==24.0
platformdirs==4.2.2
prompt_toolkit==3.0.47
protobuf==4.25.3
protoc-gen-openapiv2==0.0.1
psutil==5.9.0
ptyprocess==0.7.0
pure_eval==0.2.3
pyarrow==17.0.0
pycparser==2.22
pydantic==2.8.2
pydantic_core==2.20.1
pydeck==0.9.1
Pygments==2.18.0
PyMuPDF==1.24.8
PyMuPDFb==1.24.8
pypdf==4.3.1
pypdfium2==4.30.0
python-dateutil==2.9.0
python-dotenv==1.0.1
python-iso639==2024.4.27
python-magic==0.4.27
pytz==2024.1
PyYAML==6.0.1
pyzmq==25.1.2
rank-bm25==0.2.2
rapidfuzz==3.9.4
referencing==0.35.1
regex==2024.5.15
requests==2.32.3
requests-toolbelt==1.0.0
rich==13.7.1
rpds-py==0.19.0
setuptools==69.5.1
six==1.16.0
smmap==5.0.1
sniffio==1.3.1
soupsieve==2.5
SQLAlchemy==2.0.31
stack-data==0.6.2
streamlit==1.36.0
tabulate==0.9.0
tenacity==8.5.0
tiktoken==0.7.0
toml==0.10.2
toolz==0.12.1
tornado==6.4.1
tqdm==4.66.4
traitlets==5.14.3
types-requests==2.32.0.20240712
typing_extensions==4.12.2
typing-inspect==0.9.0
tzdata==2024.1
unstructured==0.15.0
unstructured-client==0.24.1
urllib3==2.2.2
wcwidth==0.2.13
wget==3.2
wheel==0.43.0
wrapt==1.16.0
yarl==1.9.4
zipp==3.19.2
